+++ perun/utils/exceptions.py
@@ -75,6 +75,30 @@ class EntryNotFoundException(Exception):
         return msg + " found in the index{}".format(": " + self.cause if self.cause else '')
 
 
+class IndexNotFoundException(Exception):
+    """Raised when the index file for the minor version does not exist"""
+    def __init__(self, minor_version):
+        """
+        :param str minor_version: the minor version that was supposed to have an index file
+        """
+        super().__init__("")
+        self.minor_version = minor_version
+
+    def __str__(self):
+        return "Index file for the minor version '{}' was not found.".format(self.minor_version)
+
+
+class StatsFileNotFoundException(Exception):
+    """Raised when the looked up stats file does not exist"""
+    def __init__(self, filename):
+        super().__init__("")
+        self.path = filename
+        self.msg = "The requested stats file '{}' does not exist".format(self.path)
+
+    def __str__(self):
+        return self.msg
+
+
 class VersionControlSystemException(Exception):
     """Raised when there is an issue with wrapped version control system.